How To Print More GridView Pages In C# - Basic CRUD Application MS Access Database
In this scenario, the scroll bar works when loading more pages, so the user can review all data the pages. However, I could not find how to make the print preview shows them all, or print them when you click on print. It shows only the first page in the print preview, and the same thing when printing them out. So the belows step show how to print more GridView pages in C#.